+++ tests/test_models.py
@@ -9,7 +9,10 @@
 
 import pytest
 
-from django_markov.models import _get_corpus_char_limit
+from django_markov.models import (
+    MarkovTextModel,
+    _get_corpus_char_limit,
+)
 
 pytestmark = pytest.mark.django_db(transaction=True)
 
@@ -26,6 +29,30 @@ def test_get_char_limit(
 ) -> None:
     settings.MARKOV_CORPUS_MAX_CHAR_LIMIT = override_value
     assert _get_corpus_char_limit() == expected_result
+
+
+def test_text_models_return_none_on_empty_directive():
+    model = MarkovTextModel.objects.create()
+    assert not model._as_text_model()
+    assert not model._compiled_model
+
+
+def test_do_not_recompile_compiled_model(compiled_model):
+    assert (
+        compiled_model._as_text_model().to_json()
+        == compiled_model._compiled_model.to_json()
+    )
+
+
+def test_compile_non_compiled_model(sample_corpus):
+    model = MarkovTextModel.objects.create()
+    model.update_model_from_corpus(
+        [sample_corpus, "My name is Inigo Montoya."], store_compiled=False
+    )
+    assert model.data
+    model.refresh_from_db()
+    assert not model._as_text_model().chain.compiled
+    assert model._compiled_model.chain.compiled
 
 
 @pytest.mark.asyncio
@@ -50,11 +77,11 @@ async def test_update_corpus_excessive(
     if expect_error:
         with pytest.raises(ValueError):
             await text_model.aupdate_model_from_corpus(
-                [sample_corpus, "My name is Inigo Montoya."], char_limit
+                [sample_corpus, "My name is Inigo Montoya."], char_limit=char_limit
             )
     else:
         await text_model.aupdate_model_from_corpus(
-            [sample_corpus, "My name is Inigo Montoya."], char_limit
+            [sample_corpus, "My name is Inigo Montoya."], char_limit=char_limit
         )
         await text_model.arefresh_from_db()
         assert text_model.data
@@ -75,7 +102,7 @@ async def test_update_corpus(text_model, sample_corpus) -> None:
 @pytest.mark.asyncio
 @pytest.mark.parametrize("char_limit", [0, 5000, 280, 500])
 async def test_generate_sentence(compiled_model, char_limit: int) -> None:
-    sentence = await compiled_model.agenerate_sentence(char_limit)
+    sentence = await compiled_model.agenerate_sentence(char_limit, tries=200)
     assert sentence
     if char_limit != 0:
         assert len(sentence) < char_limit